Split up dev, gui and joystick into blueprints
